Reogranize items storage tests
