+++ tests/ItemsStorageTest.php
@@ -62,230 +62,6 @@ final class ItemsStorageTest extends TestCase
         $this->assertFileExists($directory . '/items.php');
     }
 
-    public function testGetAllWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAll());
-    }
-
-    public function testGetByNamesWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getByNames(['posts.view']));
-    }
-
-    public function testGetWithConcurrency(): void
-    {
-        $this->assertNotNull($this->getEmptyConcurrentItemsStorage()->get('posts.view'));
-    }
-
-    public function testExistsWithConcurrency(): void
-    {
-        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->exists('posts.view'));
-    }
-
-    public function testRoleExistsWithConcurrency(): void
-    {
-        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->roleExists('posts.viewer'));
-    }
-
-    public function testAddWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->add(new Permission('test1'));
-        $count = count($actionStorage->getAll());
-        $actionStorage->add(new Permission('test2'));
-
-        $testStorage->add(new Permission('test1'));
-        $this->assertCount($count, $testStorage->getAll());
-    }
-
-    public function testUpdateWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $commonUpdatedItem = $actionStorage->get('posts.view')->withName('posts.view1');
-        $actionStorage->update('posts.view', $commonUpdatedItem);
-        $actionStorage->update('posts.create', $actionStorage->get('posts.create')->withName('posts.create1'));
-
-        $testStorage->update('posts.view', $commonUpdatedItem);
-        $this->assertTrue($testStorage->exists('posts.view1'));
-        $this->assertFalse($testStorage->exists('posts.create1'));
-    }
-
-    public function testRemoveWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->remove('posts.view');
-        $count = count($actionStorage->getAll());
-        $actionStorage->remove('posts.create');
-
-        $testStorage->remove('posts.view');
-        $this->assertCount($count, $testStorage->getAll());
-    }
-
-    public function testGetRolesWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getRoles());
-    }
-
-    public function testGetRolesByNamesWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getRolesByNames(['posts.viewer']));
-    }
-
-    public function testGetRoleWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getRole('posts.viewer'));
-    }
-
-    public function testClearRolesWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $permissionsCount = count($actionStorage->getPermissions());
-        $actionStorage->add(new Permission('test1'));
-
-        $testStorage->clearRoles();
-
-        $all = $testStorage->getAll();
-        $this->assertCount($permissionsCount, $all);
-        $this->assertContainsOnlyInstancesOf(Permission::class, $all);
-    }
-
-    public function testGetPermissionsWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getPermissions());
-    }
-
-    public function testGetPermissionsByNamesWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getPermissionsByNames(['posts.view']));
-    }
-
-    public function testGetPermissionWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getPermission('posts.view'));
-    }
-
-    public function testClearPermissionsWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $rolesCount = count($actionStorage->getRoles());
-        $actionStorage->add(new Role('test1'));
-
-        $testStorage->clearPermissions();
-
-        $all = $testStorage->getAll();
-        $this->assertCount($rolesCount, $all);
-        $this->assertContainsOnlyInstancesOf(Role::class, $all);
-    }
-
-    public function testGetParentsWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getParents('posts.view'));
-    }
-
-    public function testGetAccessTreeWithConcurrency(): void
-    {
-        $this->markTestSkipped();
-
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAccessTree('posts.view'));
-    }
-
-    public function testGetDirectChildrenWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getDirectChildren('posts.viewer'));
-    }
-
-    public function testGetAllChildrenWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildren('posts.viewer'));
-    }
-
-    public function testGetAllChildRolesWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildRoles('posts.redactor'));
-    }
-
-    public function testGetAllChildPermissionsWithConcurrency(): void
-    {
-        $this->assertNotEmpty($this->getEmptyConcurrentItemsStorage()->getAllChildPermissions('posts.viewer'));
-    }
-
-    public function testHasChildrenWithConcurrency(): void
-    {
-        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->hasChildren('posts.viewer'));
-    }
-
-    public function testHasChildWithConcurrency(): void
-    {
-        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->hasChild('posts.viewer', 'posts.view'));
-    }
-
-    public function testHasDirectChildWithConcurrency(): void
-    {
-        $this->assertTrue($this->getEmptyConcurrentItemsStorage()->hasDirectChild('posts.viewer', 'posts.view'));
-    }
-
-    public function testAddChildWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->addChild('posts.viewer', 'posts.create');
-        $actionStorage->addChild('posts.viewer', 'posts.update');
-
-        $testStorage->addChild('posts.viewer', 'posts.create');
-        $this->assertTrue($testStorage->hasChild('posts.viewer', 'posts.create'));
-        $this->assertFalse($testStorage->hasChild('posts.viewer', 'posts.update'));
-    }
-
-    public function testRemoveChildWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->removeChild(parentName: 'posts.redactor', childName: 'posts.create');
-        $actionStorage->removeChild(parentName: 'posts.redactor', childName: 'posts.update');
-
-        $testStorage->removeChild(parentName: 'posts.redactor', childName: 'posts.create');
-        $this->assertFalse($testStorage->hasChild(parentName: 'posts.redactor', childName: 'posts.create'));
-        $this->assertTrue($testStorage->hasChild(parentName: 'posts.redactor', childName: 'posts.update'));
-    }
-
-    public function testRemoveChildrenWithConcurrency(): void
-    {
-        $testStorage = new ItemsStorage($this->getDataPath());
-        $actionStorage = $this->getItemsStorage();
-
-        $actionStorage->removeChildren('posts.viewer');
-        $actionStorage->removeChildren('posts.redactor');
-
-        $testStorage->removeChildren('posts.viewer');
-        $this->assertFalse($testStorage->hasChildren('posts.viewer'));
-        $this->assertTrue($testStorage->hasChildren('posts.redactor'));
-    }
-
-    public function testMultipleCallsWithConcurrency(): void
-    {
-        $testStorage = $this->getItemsStorageForModificationAssertions();
-        $actionStorage = $this->getItemsStorage();
-        $count = count($actionStorage->getAll());
-
-        $actionStorage->add(new Permission('test1'));
-        $this->assertCount($count, $testStorage->getAll());
-
-        $actionStorage->add(new Permission('test2'));
-        $this->assertCount($count, $testStorage->getAll());
-    }
-
     protected function createItemsStorage(): ItemsStorageInterface
     {
         return new ItemsStorage($this->getDataPath());
@@ -293,15 +69,7 @@ final class ItemsStorageTest extends TestCase
 
     protected function getItemsStorageForModificationAssertions(): ItemsStorageInterface
     {
-        return str_ends_with($this->name(), 'WithConcurrency') ? $this->createItemsStorage() : $this->getItemsStorage();
-    }
-
-    private function getEmptyConcurrentItemsStorage(): ItemsStorageInterface
-    {
-        $storage = $this->getItemsStorageForModificationAssertions();
-        $this->getItemsStorage()->clear();
-
-        return $storage;
+        return $this->getItemsStorage();
     }
 
     private function getTempDirectory(): string
